requirements.txt
# Web
beautifulsoup4==4.12.3
websocket_client==1.8.0

# Tool
pandas==2.2.3
PyYAML==6.0.1
Pillow==11.0.0
markdown_to_json==2.1.2
opencv_python==4.10.0.84

# LLM
openai==1.54.5
langchain_core==0.3.19
langchain_openai==0.2.9
langchain_chroma==0.1.4
